requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
mysql-connector-python==9.2.0
packaging==24.2
pillow==11.1.0
pycparser==2.22
pyperclip==1.9.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==76.0.0
urllib3==2.3.0
wheel==0.45.1
